Refactor Detalleventa, Venta Producto: separate read/write logic and add new classes with its methods

+++ sistemaInventarios/src/main/java/sl/sistemaInventarios/servicio/usuario/clases/VendedorConsultaServicio.java
@@ -13,7 +13,7 @@ public class VendedorConsultaServicio implements IVendedorConsultaServicio {
     private VendedorRepositorio vendedorRepositorio;
 
     @Override
-    public Vendedor buscarVendedorPorId(Long idVendedor) {
+    public Vendedor buscarVendedorPorId(Integer idVendedor) {
         return this.vendedorRepositorio.findById(idVendedor).orElse(null);
     }
